vyzio_ads/requirements.txt
# Django & REST Framework
django>=4.2.26
djangorestframework>=3.15.2
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework-simplejwt>=5.5.1

# Database
psycopg2-binary==2.9.9

# Configuration
python-decouple==3.8

# Authentication
PyJWT>=2.8.0

# Media & Files
Pillow>=10.3.0
cloudinary==1.36.0
python-multipart>=0.0.19

# Payments
stripe==7.4.0

# Async & Cache
celery==5.3.4
redis==5.0.1

# WebSockets
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Production
gunicorn>=22.0.0
whitenoise==6.6.0
python-dateutil==2.8.2

# Monitoring (optional)
sentry-sdk>=2.8.0

# ===== DEVELOPMENT & TESTING =====
# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Linting & Formatting
black>=24.3.0
isort==5.13.2
flake8==7.0.0
bandit>=1.7.7

# Pre-commit
pre-commit==3.6.0

# Type checking (optional)
mypy==1.8.0
django-stubs>=5.2.0
